+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changes
 ----------------
 * [UI]: Fixed bug where `.xls` file could not be uploaded through the file picker on the metadata upload page. (0.21.1)
 * [UI]: Moved the position of the notification system to top center.
+* [UI]: Users can save analysis results to samples after pipeline is done in "Share Results" tab.
 
 0.20.0 to 0.21.0
 ----------------